requirements.txt
loguru==0.6.0
Pillow==9.5.0
PyBluez==0.23
